SCRUM-47 feat: create alls views of forgot password ok
